pyproject.toml
[project]
name = "slot-booking-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Anubhav Shukla",email = "anubhav008shukla@gmail.com"}
]
license = {text = "MIT"}
readme = "README.rst"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-filter (>=24.3,<25.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-sendgrid-v5 (>=1.2.4,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    {include = "slot_booking_backend"}
]

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = false
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true
indent_width = 4
continuation_indent_width = 4


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
colorlog = "^6.9.0"
